Teach duplicacy how to skip offline files
Windows 10 Fall Creator Update (version 1709) expanded the concept of offline
files to include the ability for files to be recalled on open and or data access. Whilst
the files for most purposes appear to be present on the local filesystem it is actually
a placeholder with the correct metadata. The file content is downloaded on open or
data access. Early adopters of this capability are Microsoft OneDrive and Microsoft
OneDrive for Business.

This commit teaches duplicacy a new command line option,
-skip-offline-data. For backwards compatibility, this option defaults to
false. When set to true, on Windows, any file that has the any of the
file attributes
FILE_ATTRIBUTE_OFFLINE
FILE_ATTRIBUTE_RECALL_ON_DATA_ACCESS
FILE_ATTRIBUTE_RECALL_ON_OPEN
is skipped.

Prior to this commit duplicacy would fail with an error when encountring an offline
file and
1. OneDrive client was not running
2. The file was part of OneDrive for a different user
3. The OneDrive user was not logged on.
Scenarios 2 and 3 are common when running duplicacy from Task Scheduler to
backup a multi-user PC where each user uses OneDrive.